+++ commands/ping.js
@@ -1,6 +1,6 @@
 "use strict";
 Object.defineProperty(exports, "__esModule", { value: true });
-const Command_1 = require("../Command");
+const Command_1 = require("../modules/Command");
 const names = ["ping"];
 const func = async (msg, data) => {
     await data.bot.createMessage(msg.channel.id, "Pong!");
